inline column functions inside interpreter
